+++ src/components/POLine/index.js
@@ -1,3 +1,4 @@
+export * from './ConfirmReceivingWorkflowChange';
 export { default as POLine } from './POLine';
 export { default as POLineForm } from './POLineForm';
 export { default as POLineView } from './POLineView';