+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * Add `OCLC` number to list of valid product identifier types. Refs UIOR-1440.
 * Format the "Publishers" column to avoid displaying nullish values. Fixes UIOR-1436.
 * Display an error toast only when PO Line locations violate fund restrictions. Fixes UIOR-1420.
+* Automatically set to `Independent order and receipt quantity` if a user selects `Receipt not required` for opened order. UIOR-1243.
 
 ## [8.0.3](https://github.com/folio-org/ui-orders/tree/v8.0.2) (2025-04-10)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v8.0.2...v8.0.3)